Complete the hire method
Completed the hire method for the American Airlines coding interview. In creating this method I implemented the use of iterators and a helper method in traverseInnerSets. The code works by implementing a comparison to the printOrganization method to search for a substring that determines if the position title exists within the node's children, if it does not exist within that node's children, then the iterator would iterate into the next node on the same level. If the title exists within the node's children then the recursive loop will continue until the position title is found by the iterator.